service(processor): implement real-time monitoring and filtering logic
- Loads whale whitelist from KeyDB into an in-memory map for high-speed lookups.
- Establishes a persistent WebSocket connection to the Arbitrum node and subscribes to new pending transactions.
- For each received transaction hash, fetches full details (from, to) via an HTTP RPC call.
- Filters transactions against the whitelist and logs detected whale activity.
